fix: Improved support for pickle adding tests for listeners and custom instance values
